+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Localized the content of the about page
 - Refactored the dialog footer component
 - Refactored the dialog header component
+- Refactored the account detail dialog component to standalone
 - Refactored the benchmark comparator component to standalone
 - Refactored the portfolio summary component to standalone
 - Refactored the world map chart component to standalone